+++ src/main/java/lotto/controller/LottoController.java
@@ -1,5 +1,9 @@
 package lotto.controller;
 
+import lotto.controller.input.InputController;
+import lotto.controller.logic.LogicController;
+import lotto.controller.output.OutputController;
+
 public class LottoController {
 
     public LottoController() {
