+++ src/main/java/lotto/controller/input/InputController.java
@@ -1,0 +1,10 @@
+package lotto.controller.input;
+
+public class InputController {
+
+    public InputController() {
+        new InputMoneyController();
+        new InputWinningNumberController();
+        new InputBonusController();
+    }
+}
