+++ src/main/java/lotto/controller/output/OutputController.java
@@ -1,0 +1,17 @@
+package lotto.controller.output;
+
+import lotto.domain.caculate.CalculateLottoResult;
+import lotto.domain.fomatting.PrizeFormatter;
+import lotto.domain.fomatting.RateFormatter;
+import lotto.view.output.OutputWinningCount;
+import lotto.view.output.OutputWinningRate;
+
+public class OutputController {
+
+    public OutputController() {
+        OutputWinningCount.printWinningIntro();
+        OutputWinningCount.printWinningCount(PrizeFormatter.getPrize(),
+            CalculateLottoResult.getLottoResult());
+        OutputWinningRate.printRate(RateFormatter.getRate());
+    }
+}
